Move relevance checkers to util file; Rename test file
